Functional component add useEffect clean up code
While doing unit tesing identified this issue. Warning: Can't perform a React state update on
an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.

Github issue id: #212
Testing done : Wrote clean_up code for the functional components which has useEffect.